feat: move product package
