+++ src/main/java/kitchenpos/domain/product/Product.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.product;
 
 import kitchenpos.domain.vo.Money;
