+++ src/test/java/kitchenpos/domain/menu/MenuValidatorTest.java
@@ -1,7 +1,7 @@
 package kitchenpos.domain.menu;
 
-import kitchenpos.domain.Product;
-import kitchenpos.domain.ProductRepository;
+import kitchenpos.domain.product.Product;
+import kitchenpos.domain.product.ProductRepository;
 import kitchenpos.support.ServiceTest;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
