+++ src/test/java/kitchenpos/domain/product/ProductTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.product;
 
 import org.junit.jupiter.api.Test;
