+++ src/test/java/kitchenpos/fixture/ProductFixture.java
@@ -1,7 +1,7 @@
 package kitchenpos.fixture;
 
 import kitchenpos.application.dto.ProductRequest;
-import kitchenpos.domain.Product;
+import kitchenpos.domain.product.Product;
 
 import java.math.BigDecimal;
